+++ changelog.md
@@ -8,6 +8,10 @@ https://www.dropbox.com/sh/4ctl2y928xkak4f/AAADEZj_hFpGQaNOdd3yqcAHa?dl=0
 
 https://1drv.ms/f/s!ArwKS_ZUR01gySQIE1QCL2i9F_uR
 
+- new: added ConvertFromDoubleWidth to AviSynth profiles
+
+- update: ffmpeg, unlike the built before this one supports mp3 encoding
+
 ### 2.0.1.2
 
 - fix for audio streams demuxed as mka